Improve performance and logic around keyword collapsing
The original keyword collpasing algorithm was horribly non-performant
and rather dumb.  The new logic is around 20 times faster in most of my
tests, and it tries to be more intelligent about when to replace single
word keywords with multi-words (vs just adding the multi-words and
keeping one or more of the singles).  Still room for improvement, but
hopefully this is more production quality.