Fix nil can't be coerced into Float error
When ASCII folding is turned on in addition to the token collapser,
there was a chance that a non-ASCII token might trigger a "nil can't be
coerced into Float" error.  This is because the regex matches the
original text (containing the non-ASCII text) even though the token
permutation is ASCII folded.  In this scenario the best thing to do is
to ignore the error and proceed forward as best we can.