Allow arrays in extract
When processing multiple documents it may not be advisable to
concatenate the text together into one single large piece of text.  This
is because the tokens on the boundary (those at the beginning/end of
each document) will have graph edges added to tokens accross the
boundary into the other document, creating a false contextual
relationship between them.

But because it may be desirable to pass multiple documents into the
extractor to obtain a unified ranking across the documents, we will add
support for passing in an array of text into the extractor which will
provide slightly different values than if the concatenated text were
passed in.