Fix segmentation fault caused by GC
Use of rb_hash_new() to create a node lookup was reckless.  As soon as
the garbage collector runs this can get blown away and result in a seg
fault.

The fix is a little hacky, but it seems to be the best I can do since
there isn't a great way that I can see to initialize a plain hash in C
that accepts any ruby object as a key.  Those internal methods are
hidden inside ruby.c and not exposed.  `st_init_table_with_size` is the
method that initializes a new hash table, but it requires passing in
information about how to compare/hash the keys of the data going into
it.  There's a lot of code in ruby.c that I would have to copy over to
get the behavior we need, but getting it from
`rb_hash_tbl(rb_hash_new())->type` seems relatively safe given that it
should just be a reference to a global function definition in the ruby
library.